+++ packages/RSuite/NEWS.md
@@ -1,6 +1,7 @@
 # RSuite 0.36 (2019-03-??)
   * #184: detecting declaration of dependencies in scripts with "library (...)".
   * handling of branches in Git RC adapter fixed.
+  * #183: local Git repository is created while starting project if not RC detected 
   
 # RSuite 0.35 (2019-01-07)
   * Fix in dependecy resolving process. Upper requirements enforse subdependencies